move kb resolution into accessibleViewService
