+++ src/vs/workbench/contrib/accessibility/browser/extensionAccesibilityHelp.contribution.ts
@@ -3,17 +3,14 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { MarkdownString } from 'vs/base/common/htmlContent';
 import { DisposableMap, IDisposable, DisposableStore, Disposable } from 'vs/base/common/lifecycle';
+import { ServicesAccessor } from 'vs/editor/browser/editorExtensions';
 import { AccessibleViewType, ExtensionContentProvider } from 'vs/platform/accessibility/browser/accessibleView';
 import { AccessibleViewRegistry } from 'vs/platform/accessibility/browser/accessibleViewRegistry';
-import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
-import { IPickerQuickAccessItem } from 'vs/platform/quickinput/browser/pickerQuickAccess';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { FocusedViewContext } from 'vs/workbench/common/contextkeys';
 import { IViewsRegistry, Extensions, IViewDescriptor } from 'vs/workbench/common/views';
-import { AccessibilityCommandId } from 'vs/workbench/contrib/accessibility/common/accessibilityCommands';
 import { IViewsService } from 'vs/workbench/services/views/common/viewsService';
 
 export class ExtensionAccessibilityHelpDialogContribution extends Disposable {
@@ -42,9 +39,9 @@ export class ExtensionAccessibilityHelpDialogContribution extends Disposable {
 
 function registerAccessibilityHelpAction(keybindingService: IKeybindingService, viewDescriptor: IViewDescriptor): IDisposable {
 	const disposableStore = new DisposableStore();
-	const helpContent = resolveExtensionHelpContent(keybindingService, viewDescriptor.accessibilityHelpContent);
-	if (!helpContent) {
-		throw new Error('No help content for view');
+	const content = viewDescriptor.accessibilityHelpContent?.value;
+	if (!content) {
+		throw new Error('No content provided for the accessibility help dialog');
 	}
 	disposableStore.add(AccessibleViewRegistry.register({
 		priority: 95,
@@ -55,8 +52,8 @@ function registerAccessibilityHelpAction(keybindingService: IKeybindingService, 
 			const viewsService = accessor.get(IViewsService);
 			return new ExtensionContentProvider(
 				viewDescriptor.id,
-				{ type: AccessibleViewType.Help, configureKeybindingItems: helpContent.configureKeybindingItems },
-				() => helpContent.value.value,
+				{ type: AccessibleViewType.Help },
+				() => content,
 				() => viewsService.openView(viewDescriptor.id, true),
 			);
 		}
@@ -68,35 +65,3 @@ function registerAccessibilityHelpAction(keybindingService: IKeybindingService, 
 	}));
 	return disposableStore;
 }
-
-function resolveExtensionHelpContent(keybindingService: IKeybindingService, content?: MarkdownString): { value: MarkdownString; configureKeybindingItems: IPickerQuickAccessItem[] | undefined } | undefined {
-	if (!content) {
-		return;
-	}
-	const configureKeybindingItems: IPickerQuickAccessItem[] = [];
-	let resolvedContent = typeof content === 'string' ? content : content.value;
-	const matches = resolvedContent.matchAll(/\<keybinding:(?<commandId>.*)\>/gm);
-	for (const match of [...matches]) {
-		const commandId = match?.groups?.commandId;
-		let kbLabel;
-		if (match?.length && commandId) {
-			const keybinding = keybindingService.lookupKeybinding(commandId)?.getAriaLabel();
-			if (!keybinding) {
-				const configureKb = keybindingService.lookupKeybinding(AccessibilityCommandId.AccessibilityHelpConfigureKeybindings)?.getAriaLabel();
-				const keybindingToConfigureQuickPick = configureKb ? '(' + configureKb + ')' : 'by assigning a keybinding to the command accessibility.openQuickPick.';
-				kbLabel = `, configure a keybinding ` + keybindingToConfigureQuickPick;
-				configureKeybindingItems.push({
-					label: commandId,
-					id: commandId
-				});
-			} else {
-				kbLabel = ' (' + keybinding + ')';
-			}
-			resolvedContent = resolvedContent.replace(match[0], kbLabel);
-		}
-	}
-	const value = new MarkdownString(resolvedContent);
-	value.isTrusted = true;
-	return { value, configureKeybindingItems: configureKeybindingItems.length ? configureKeybindingItems : undefined };
-}
-
